chore: デバッグ用 ItemSlotView を Client.DebugSystem へ移し Starter の uGUI 事前ロードを外す
ItemSelectModal だけが使うスロット表示を Client.DebugSystem/ItemSlot へ移し、
ツールチップ・スロット装飾など残る呼び出し元が無い API を落とした。
起動時の PreloadCriticalAssetsAsync（ChestBlockInventory / FluidSlotView 前提）は
prefab ごと消えるため廃し、prefab のロードはデバッグモーダル自身が行う。
