+++ moorestech_client/Assets/Scripts/Client.DebugSystem/ItemSlot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2407f25bbcc7f4024a05dc69d1a0d229
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
