+++ moorestech_client/Assets/Scripts/Client.DebugSystem/ItemSlot/CommonSlotView.cs
@@ -1,33 +1,22 @@
-// [uGUI廃止Phase1] uGUI描画は恒久停止・ビューは未メンテ。ただし本クラスは外部（Web UIブリッジ等）から参照中のため削除前に整理が必要（docs/webui/ugui-retirement-plan.md）
-// [uGUI retirement Phase1] uGUI rendering is permanently disabled and the view is unmaintained, but this class is still referenced externally (e.g. Web UI bridge); untangle before deletion (docs/webui/ugui-retirement-plan.md)
 using System;
-using Core.Master;
 using TMPro;
 using UniRx;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
-namespace Client.Game.InGame.UI.Inventory.Common
+namespace Client.DebugSystem.ItemSlot
 {
+    /// <summary>
+    ///     デバッグUI（ItemSelectModal）専用のスロット描画とポインタ入力
+    ///     Slot rendering and pointer input used only by the debug UI (ItemSelectModal)
+    /// </summary>
     public class CommonSlotView : MonoBehaviour, IPointerDownHandler, IPointerUpHandler, IPointerEnterHandler, IPointerClickHandler, IPointerExitHandler, IPointerMoveHandler
     {
-        [SerializeField] private RectTransform rootRectTransform;
-        
         [SerializeField] private Image itemImage;
         
-        [SerializeField] private GameObject normalFrame;
-        [SerializeField] private GameObject machineSlotFrame;
-        [SerializeField] private GameObject craftRecipeFrame;
-        
-        [SerializeField] private GameObject hotBarSelect;
-        
-        [SerializeField] private GameObject grayOutImage;
-        [SerializeField] private GameObject hoverImage; // TODO 後で対応
-        [SerializeField] private GameObject clickImage; // TODO 後で対応
-        
-        [SerializeField] private GameObject normalItemSlotObject;
-        [SerializeField] private GameObject noneCrossObject;
+        [SerializeField] private GameObject hoverImage;
+        [SerializeField] private GameObject clickImage;
         
         [SerializeField] private TMP_Text countText;
         
@@ -50,7 +39,6 @@ namespace Client.Game.InGame.UI.Inventory.Common
             _onLeftClickUp.Subscribe(_ => clickImage.SetActive(false)).AddTo(this);
         }
         
-        
         public void SetView(Sprite sprite, string count)
         {
             countText.text = count;
@@ -59,57 +47,10 @@ namespace Client.Game.InGame.UI.Inventory.Common
             itemImage.sprite = sprite;
         }
         
-        public void SetCountText(string count)
-        {
-            countText.text = count;
-        }
-
         public void SetViewClear()
         {
             countText.text = string.Empty;
             itemImage.gameObject.SetActive(false);
-        }
-        
-        public void SetSlotViewOption(CommonSlotViewOption slotOption)
-        {
-            if (slotOption.GrayOut != null) SetGrayOut(slotOption.GrayOut.Value);
-            if (slotOption.HotBarSelected != null) SetHotBarSelect(slotOption.HotBarSelected.Value);
-            if (slotOption.ItemSlotFrameType != null) SetFrame(slotOption.ItemSlotFrameType.Value);
-            if (slotOption.ItemSlotType != null) SetItemSlotType(slotOption.ItemSlotType.Value);
-            if (slotOption.CountTextFontSize != null) countText.fontSize = slotOption.CountTextFontSize.Value; 
-            if (slotOption.SizeDelta != null) rootRectTransform.sizeDelta = slotOption.SizeDelta.Value;
-            
-            #region Internal
-            
-            void SetFrame(ItemSlotFrameType frameType)
-            {
-                normalFrame.SetActive(frameType == ItemSlotFrameType.Normal);
-                machineSlotFrame.SetActive(frameType == ItemSlotFrameType.MachineSlot);
-                craftRecipeFrame.SetActive(frameType == ItemSlotFrameType.CraftRecipe);
-            }
-            
-            void SetGrayOut(bool active)
-            {
-                grayOutImage.SetActive(active);
-            }
-            
-            void SetItemSlotType(ItemSlotType slotType)
-            {
-                normalItemSlotObject.SetActive(slotType == ItemSlotType.Normal);
-                noneCrossObject.SetActive(slotType == ItemSlotType.NoneCross);
-            }
-            
-            void SetHotBarSelect(bool active)
-            {
-                hotBarSelect.SetActive(active);
-            }
-            
-            #endregion
-        }
-        
-        public void SetActive(bool active)
-        {
-            gameObject.SetActive(active);
         }
         
         
@@ -118,21 +59,14 @@ namespace Client.Game.InGame.UI.Inventory.Common
         public IObservable<(CommonSlotView, ItemUIEventType)> OnPointerEvent => _onPointerEvent;
         private readonly Subject<(CommonSlotView, ItemUIEventType)> _onPointerEvent = new();
         
-        public IObservable<CommonSlotView> OnRightClickDown => _onRightClickDown;
-        private readonly Subject<CommonSlotView> _onRightClickDown = new();
-        public IObservable<CommonSlotView> OnLeftClickDown => _onLeftClickDown;
-        private readonly Subject<CommonSlotView> _onLeftClickDown = new();
         public IObservable<CommonSlotView> OnRightClickUp => _onRightClickUp;
         private readonly Subject<CommonSlotView> _onRightClickUp = new();
-        public IObservable<CommonSlotView> OnLeftClickUp => _onLeftClickUp;
+        private readonly Subject<CommonSlotView> _onRightClickDown = new();
+        private readonly Subject<CommonSlotView> _onLeftClickDown = new();
         private readonly Subject<CommonSlotView> _onLeftClickUp = new();
-        public IObservable<CommonSlotView> OnCursorEnter => _onCursorEnter;
         private readonly Subject<CommonSlotView> _onCursorEnter = new();
-        public IObservable<CommonSlotView> OnCursorExit => _onCursorExit;
         private readonly Subject<CommonSlotView> _onCursorExit = new();
-        public IObservable<CommonSlotView> OnCursorMove => _onCursorMove;
         private readonly Subject<CommonSlotView> _onCursorMove = new();
-        public IObservable<CommonSlotView> OnDoubleClick => _onDoubleClick;
         private readonly Subject<CommonSlotView> _onDoubleClick = new();
         
         private void OnInvokeOtherEvent((CommonSlotView, ItemUIEventType) data)
