+++ moorestech_client/Assets/Scripts/Client.DebugSystem/ItemSlot/CommonSlotView.cs.meta
