+++ moorestech_client/Assets/Scripts/Client.DebugSystem/ItemSlot/ItemUIEventType.cs
@@ -1,0 +1,16 @@
+namespace Client.DebugSystem.ItemSlot
+{
+    public enum ItemUIEventType
+    {
+        RightClickDown,
+        LeftClickDown,
+        RightClickUp,
+        LeftClickUp,
+        
+        CursorEnter,
+        CursorExit,
+        CursorMove,
+        
+        DoubleClick,
+    }
+}
