+++ moorestech_client/Assets/Scripts/Client.DebugSystem/ItemSlot/ItemUIEventType.cs.meta
