--- moorestech_client/Assets/Scripts/Client.Game/InGame/UI/Inventory/Common.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 0f4ad4ad7ff84ed39da305857677dc2e
-timeCreated: 1643812958
