--- moorestech_client/Assets/Scripts/Client.Game/InGame/UI/Inventory/Common/CommonSlotViewExtension.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: a6d15322b0bc43b9acda8f8f62422592
-timeCreated: 1749982080
