--- moorestech_client/Assets/Scripts/Client.Game/InGame/UI/Inventory/Common/CommonSlotViewOption.cs
@@ -1,47 +1,0 @@
-// [uGUI廃止Phase1] Web UI移行済みのため未メンテ・描画恒久停止。Phase2で削除予定（docs/webui/ugui-retirement-plan.md）
-// [uGUI retirement Phase1] Unmaintained; rendering permanently disabled after the Web UI migration. Slated for deletion in Phase2 (docs/webui/ugui-retirement-plan.md)
-using UnityEngine;
-
-namespace Client.Game.InGame.UI.Inventory.Common
-{
-    /// <summary>
-    /// スロットの設定を変更するクラス。非nullの値のみ変更を実施する
-    /// A class to change slot settings. Only non-null values are changed.
-    /// </summary>
-    public class CommonSlotViewOption
-    {
-        public bool? GrayOut;
-        public ItemSlotFrameType? ItemSlotFrameType;
-        public ItemSlotType? ItemSlotType;
-        public bool? HotBarSelected;
-        public float? CountTextFontSize;
-        public Vector2? SizeDelta;
-    }
-    
-    public enum ItemUIEventType
-    {
-        RightClickDown,
-        LeftClickDown,
-        RightClickUp,
-        LeftClickUp,
-        
-        CursorEnter,
-        CursorExit,
-        CursorMove,
-        
-        DoubleClick,
-    }
-    
-    public enum ItemSlotType
-    {
-        Normal, // 通常のアイテム表示
-        NoneCross, // アイテムが何もないクロス表示
-    }
-    
-    public enum ItemSlotFrameType
-    {
-        Normal,
-        MachineSlot,
-        CraftRecipe,
-    }
-}
